+++ stream-chat-android-docs/src/main/kotlin/io/getstream/chat/docs/kotlin/compose/general/CustomAttachments.kt
@@ -3,8 +3,10 @@
 package io.getstream.chat.docs.kotlin.compose.general
 
 import android.os.Bundle
+import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.appcompat.app.AppCompatActivity
+import com.google.firebase.components.Component
 import io.getstream.chat.android.compose.ui.attachments.AttachmentFactory
 import io.getstream.chat.android.compose.ui.attachments.StreamAttachmentFactories
 import io.getstream.chat.android.compose.ui.theme.ChatTheme
@@ -14,7 +16,7 @@ import io.getstream.chat.android.compose.ui.theme.ChatTheme
  */
 private object CustomAttachmentsSnippet {
 
-    class MyActivity : AppCompatActivity() {
+    class CustomAttachments : ComponentActivity() {
 
         override fun onCreate(savedInstanceState: Bundle?) {
             super.onCreate(savedInstanceState)
@@ -24,7 +26,23 @@ private object CustomAttachmentsSnippet {
             setContent {
                 // override the default factories by adding your own
                 ChatTheme(attachmentFactories = myAttachmentFactories + defaultFactories) {
-                    // ChannelsScreen(...)
+                    // Chat components
+                }
+            }
+        }
+    }
+
+    class CustomQuotedAttachments: ComponentActivity(){
+
+        override fun onCreate(savedInstanceState: Bundle?) {
+            super.onCreate(savedInstanceState)
+            val myQuotedAttachmentFactories = listOf<AttachmentFactory>()
+            val defaultFactories = StreamAttachmentFactories.defaultQuotedFactories()
+
+            setContent {
+                // override the default factories by adding your own
+                ChatTheme(quotedAttachmentFactories = myQuotedAttachmentFactories + defaultFactories) {
+                    // Chat components
                 }
             }
         }